Update comments and move FromCpuStatus
