+++ src/main/java/org/edumips64/client/Result.java
@@ -1,7 +1,8 @@
 /* Result.java
  *
- * Thin value object to represent the result of a generic operation, exported
- * as a type to JavaScript.
+ * Value object to represent the result of a generic operation, exported
+ * as a type to JavaScript. Includes the full state of the CPU.
+ * 
  * (c) 2020 Andrea Spadaccini
  *
  * This file is part of the EduMIPS64 project, and is released under the GNU
