+++ src/main/java/org/edumips64/client/ResultFactory.java
@@ -1,3 +1,27 @@
+/* ResultFactory.java
+ *
+ * A factory class to generate Result objects.
+ * Injects a representation of the CPU status in every Result object that's created. 
+ * 
+ * (c) 2020 Andrea Spadaccini
+ *
+ * This file is part of the EduMIPS64 project, and is released under the GNU
+ * General Public License.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
 package org.edumips64.client;
 
 import java.util.Map;
@@ -8,16 +32,27 @@ import com.google.gwt.json.client.JSONArray;
 import org.edumips64.core.CPU;
 import org.edumips64.core.Memory;
 import org.edumips64.core.Register;
+import org.edumips64.core.CPU.CPUStatus;
 import org.edumips64.core.Pipeline.Stage;
 import org.edumips64.core.fpu.RegisterFP;
 import org.edumips64.core.is.InstructionInterface;
 
-/* Factory class to generate Result objects.
-   Injects a representation of the CPU status in every Result object that's created. */
 public class ResultFactory {
     private CPU cpu;
     private Memory memory;
     private Logger logger = Logger.getLogger("ResultFactory");
+
+    static String FromCpuStatus(CPUStatus s) {
+        switch (s) {
+        case READY:
+            return "READY";
+        case RUNNING:
+        case STOPPING:
+            return "RUNNING";
+        default:
+            return "STOPPED";
+        }
+    }
 
     public ResultFactory(CPU cpu, Memory memory) {
         this.cpu = cpu;
@@ -35,7 +70,7 @@ public class ResultFactory {
     }
 
     private Result AddCpuInfo(Result r) {
-        r.status = Simulator.FromCpuStatus(cpu.getStatus());
+        r.status = FromCpuStatus(cpu.getStatus());
         r.pipeline = getPipeline();
         r.memory = getMemory();
         r.registers = getRegisters();
